refactor: introduce domain-specific API services
