+++ frontend/src/app/companies/company.model.ts
@@ -2,7 +2,7 @@ import {
   Company as ApiCompany,
   CreateCompany as ApiCreateCompany,
   UpdateCompany as ApiUpdateCompany,
-} from '../api.service';
+} from '../api/companies-api.service';
 
 export type Company = ApiCompany;
 export type CreateCompany = ApiCreateCompany;
