+++ frontend/src/app/customers/customer.service.spec.ts
@@ -1,24 +1,24 @@
 import { TestBed } from '@angular/core/testing';
 import { of } from 'rxjs';
 import { CustomerService } from './customer.service';
-import { ApiService } from '../api.service';
+import { CustomersApiService } from '../api/customers-api.service';
 
 describe('CustomerService', () => {
   let service: CustomerService;
-  let apiSpy: jasmine.SpyObj<ApiService>;
+  let apiSpy: jasmine.SpyObj<CustomersApiService>;
 
   beforeEach(() => {
-    apiSpy = jasmine.createSpyObj<ApiService>('ApiService', ['getCustomers']);
+    apiSpy = jasmine.createSpyObj<CustomersApiService>('CustomersApiService', ['getCustomers']);
     apiSpy.getCustomers.and.returnValue(of({ items: [], total: 0 }));
 
     TestBed.configureTestingModule({
-      providers: [CustomerService, { provide: ApiService, useValue: apiSpy }],
+      providers: [CustomerService, { provide: CustomersApiService, useValue: apiSpy }],
     });
 
     service = TestBed.inject(CustomerService);
   });
 
-  it('should call ApiService.getCustomers', () => {
+  it('should call CustomersApiService.getCustomers', () => {
     service.getCustomers().subscribe();
     // eslint-disable-next-line @typescript-eslint/unbound-method
     expect(apiSpy.getCustomers).toHaveBeenCalled();
