+++ frontend/src/app/dashboard/dashboard.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, inject, signal } from '@angular/core';
 import { RouterLink } from '@angular/router';
-import { ApiService } from '../api.service';
+import { JobsApiService } from '../api/jobs-api.service';
+import { EquipmentApiService } from '../api/equipment-api.service';
+import { UsersApiService } from '../api/users-api.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -26,7 +28,9 @@ import { ApiService } from '../api.service';
   styleUrl: './dashboard.component.scss',
 })
 export class DashboardComponent implements OnInit {
-  private readonly api = inject(ApiService);
+  private readonly jobsApi = inject(JobsApiService);
+  private readonly equipmentApi = inject(EquipmentApiService);
+  private readonly usersApi = inject(UsersApiService);
 
   protected readonly upcomingJobs = signal(0);
   protected readonly equipmentAvailable = signal(0);
@@ -34,11 +38,13 @@ export class DashboardComponent implements OnInit {
   protected readonly activeUsers = signal(0);
 
   ngOnInit(): void {
-    this.api.getUpcomingJobs().subscribe((data) => this.upcomingJobs.set(data.total));
-    this.api
+    this.jobsApi.getUpcomingJobs().subscribe((data) => this.upcomingJobs.set(data.total));
+    this.equipmentApi
       .getEquipmentCount('available')
       .subscribe((data) => this.equipmentAvailable.set(data.total));
-    this.api.getEquipmentCount('in_use').subscribe((data) => this.equipmentInUse.set(data.total));
-    this.api.getUsers().subscribe((data) => this.activeUsers.set(data.length));
+    this.equipmentApi
+      .getEquipmentCount('in_use')
+      .subscribe((data) => this.equipmentInUse.set(data.total));
+    this.usersApi.getUsers().subscribe((data) => this.activeUsers.set(data.length));
   }
 }
