+++ frontend/src/app/users/user.model.ts
@@ -2,7 +2,7 @@ import {
   User as ApiUser,
   CreateUser as ApiCreateUser,
   UpdateUser as ApiUpdateUser,
-} from '../api.service';
+} from '../api/users-api.service';
 
 export type User = ApiUser;
 export type CreateUser = ApiCreateUser;
